refactor(menu): split file_lightbar.go into per-responsibility files
Pure code movement, no logic changes (top-level declaration set is
identical before/after; all menu tests pass). file_lightbar.go was 1,345
lines; now that runListFilesLightbar is a fileLightbar struct with
methods, split the methods out by responsibility:

- file_lightbar.go (831): struct, runListFilesLightbar constructor, run() loop
- file_lightbar_render.go: file-row / separator / cmd-bar / page-indicator rendering
- file_lightbar_layout.go: geometry, paging, and selection helpers
- file_lightbar_placeholders.go: @FPAGE@/@FTOTAL@/@FCONFPATH@ template
  substitution and FILE_ID.DIZ formatting

Import blocks were regenerated per file (only the symbols each file uses).
