+++ internal/menu/file_lightbar.go
@@ -8,7 +8,6 @@ import (
 	"os"
 	"path/filepath"
 	"regexp"
-	"strconv"
 	"strings"
 	"time"
 
@@ -23,72 +22,6 @@ import (
 	"github.com/ViSiON-3/vision-3-bbs/internal/user"
 	"github.com/ViSiON-3/vision-3-bbs/internal/ziplab"
 )
-
-// fileListPlaceholderRegex matches @FPAGE@, @FTOTAL@, @FCONFPATH@ with optional alignment and width.
-// Modifier: | (0x7C) or │ (CP437 0xB3, common in ANSI art) followed by L/R/C — matches message-header format.
-// Groups: 1=code (FPAGE|FTOTAL|FCONFPATH), 2=modifier (L|R|C), 3=:N digits, 4=# sequence
-var fileListPlaceholderRegex = regexp.MustCompile(`@(FPAGE|FTOTAL|FCONFPATH)(?:[\x7C\xB3]([LRC]))?(?::(\d+)|(#+))?@`)
-
-// processFileListPlaceholders replaces file-list-specific pipe codes and @-placeholders
-// with current page, total pages, total file count, and conference path. Use in FILELIST.TOP and FILELIST.BOT.
-// Pipe codes: |FPAGE ("Page X of Y"), |FTOTAL (total file count), |FCONFPATH (Conference > File Area).
-// Placeholders support alignment modifiers: @FPAGE|R###@, @FTOTAL|C:5@, @FCONFPATH|R############@
-// fconfpath is the pre-formatted "Conference > Area" string with pipe codes (from resolveFileConferencePath).
-func processFileListPlaceholders(data []byte, currentPage, totalPages, totalFiles int, fconfpath string) []byte {
-	s := string(data)
-	pageStr := fmt.Sprintf("Page %d of %d", currentPage, totalPages)
-	totalStr := strconv.Itoa(totalFiles)
-
-	// Process @-placeholders FIRST so |FPAGE inside @FPAGE|R#####@ isn't consumed by pipe codes.
-	// @CODE@ placeholders with optional alignment modifier (|L, |R, |C) and width (:N or ###)
-	// For ###: width = total placeholder length (entire token) so replacement preserves ANSI layout.
-	// E.g. @FPAGE|R###########@ is 20 cols — output is padded/truncated to 20 visible chars.
-	s = fileListPlaceholderRegex.ReplaceAllStringFunc(s, func(match string) string {
-		subs := fileListPlaceholderRegex.FindStringSubmatch(match)
-		if len(subs) < 2 {
-			return match
-		}
-		code := subs[1]
-		modifier := ""
-		if len(subs) > 2 {
-			modifier = subs[2]
-		}
-		width := 0
-		if len(subs) > 3 && subs[3] != "" {
-			width, _ = strconv.Atoi(subs[3])
-		} else if len(subs) > 4 && subs[4] != "" {
-			// Visual width: entire placeholder length (matches message-header / editor placeholder behavior)
-			width = len(match)
-		}
-		align := ansi.AlignLeft
-		if modifier != "" {
-			align = ansi.ParseAlignment(modifier)
-		}
-
-		var value string
-		switch code {
-		case "FPAGE":
-			value = pageStr
-		case "FTOTAL":
-			value = totalStr
-		case "FCONFPATH":
-			value = fconfpath
-		default:
-			return match
-		}
-		if width <= 0 {
-			return value
-		}
-		return ansi.ApplyWidthConstraintAligned(value, width, align)
-	})
-
-	// Pipe codes AFTER @-placeholders so |FPAGE inside @FPAGE|R#####@ isn't destroyed.
-	s = strings.ReplaceAll(s, "|FPAGE", pageStr)
-	s = strings.ReplaceAll(s, "|FTOTAL", totalStr)
-	s = strings.ReplaceAll(s, "|FCONFPATH", fconfpath)
-
-	return []byte(s)
-}
 
 // fileLightbar holds the state for the interactive file-area browser.
 // runListFilesLightbar builds one and drives it via run().
@@ -314,417 +247,6 @@ func runListFilesLightbar(e *MenuExecutor, s ssh.Session, terminal *term.Termina
 		cmdIndex:             cmdIndex,
 	}
 	return lb.run()
-}
-
-func (lb *fileLightbar) isFileTagged(fileID uuid.UUID) bool {
-	for _, taggedID := range lb.currentUser.TaggedFileIDs {
-		if taggedID == fileID {
-			return true
-		}
-	}
-	return false
-}
-
-func (lb *fileLightbar) stripAnsi(str string) string {
-	return lb.ansiRe.ReplaceAllString(str, "")
-}
-
-func (lb *fileLightbar) fileEntryHeight(idx int) int {
-	if idx < 0 || idx >= len(lb.allFiles) {
-		return 1
-	}
-	dizCount := len(formatDIZLines(lb.allFiles[idx].Description, lb.descColWidth, dizMaxLines))
-	if dizCount < 1 {
-		return 1
-	}
-	return dizCount
-}
-
-func (lb *fileLightbar) filesVisibleFrom(startIdx int) int {
-	usedLines := 0
-	count := 0
-	for idx := startIdx; idx < len(lb.allFiles) && usedLines < lb.visibleRows; idx++ {
-		h := lb.fileEntryHeight(idx)
-		if usedLines+1 > lb.visibleRows {
-			break
-		}
-		if usedLines+h > lb.visibleRows {
-			h = lb.visibleRows - usedLines // show as many DIZ lines as fit
-		}
-		usedLines += h
-		count++
-	}
-	return count
-}
-
-func (lb *fileLightbar) topIndexForPrevPage() int {
-	if lb.topIndex <= 0 {
-		return 0
-	}
-	usedLines := 0
-	newTop := lb.topIndex
-	for idx := lb.topIndex - 1; idx >= 0; idx-- {
-		h := lb.fileEntryHeight(idx)
-		if usedLines+h > lb.visibleRows {
-			break
-		}
-		usedLines += h
-		newTop = idx
-	}
-	return newTop
-}
-
-func (lb *fileLightbar) calculatePageInfo() (currentPage int, totalPagesCalc int) {
-	if len(lb.allFiles) == 0 {
-		return 1, 1
-	}
-	page := 0
-	idx := 0
-	foundCurrent := false
-	for idx < len(lb.allFiles) {
-		page++
-		usedLines := 0
-		pageStart := idx
-		for idx < len(lb.allFiles) && usedLines < lb.visibleRows {
-			h := lb.fileEntryHeight(idx)
-			if usedLines+1 > lb.visibleRows {
-				break
-			}
-			if usedLines+h > lb.visibleRows {
-				h = lb.visibleRows - usedLines
-			}
-			usedLines += h
-			idx++
-		}
-		if !foundCurrent && lb.topIndex >= pageStart && lb.topIndex < idx {
-			currentPage = page
-			foundCurrent = true
-		}
-	}
-	if !foundCurrent {
-		currentPage = page
-	}
-	totalPagesCalc = page
-	return currentPage, totalPagesCalc
-}
-
-func (lb *fileLightbar) clampSelection() {
-	if len(lb.allFiles) == 0 {
-		lb.selectedIndex = 0
-		lb.topIndex = 0
-		return
-	}
-	if lb.selectedIndex < 0 {
-		lb.selectedIndex = 0
-	}
-	if lb.selectedIndex >= len(lb.allFiles) {
-		lb.selectedIndex = len(lb.allFiles) - 1
-	}
-	if lb.selectedIndex < lb.topIndex {
-		lb.topIndex = lb.selectedIndex
-	}
-	// Scroll down: advance topIndex until selectedIndex fits within the
-	// visible screen area, accounting for multi-line file entries.
-	// We keep advancing until the selected entry either fits at full
-	// height or is at the very top of the viewport (so large entries
-	// like 21-line ANS art are always shown from the top, not crammed
-	// at the bottom with only a few lines visible).
-	for lb.topIndex <= lb.selectedIndex {
-		usedLines := 0
-		fits := false
-		for idx := lb.topIndex; idx < len(lb.allFiles) && usedLines < lb.visibleRows; idx++ {
-			h := lb.fileEntryHeight(idx)
-			if usedLines+1 > lb.visibleRows {
-				break // can't fit even the first line
-			}
-			fullH := h
-			if usedLines+h > lb.visibleRows {
-				h = lb.visibleRows - usedLines
-			}
-			usedLines += h
-			if idx == lb.selectedIndex {
-				// Fits fully, or entry is already at top of viewport
-				// (can't scroll any higher without losing the selection).
-				if h == fullH || idx == lb.topIndex {
-					fits = true
-				}
-				break
-			}
-		}
-		if fits || lb.topIndex >= lb.selectedIndex {
-			break
-		}
-		lb.topIndex++
-	}
-	if lb.topIndex < 0 {
-		lb.topIndex = 0
-	}
-}
-
-func (lb *fileLightbar) buildFileEntry(idx int, highlighted bool, maxLines int) []string {
-	if idx < 0 || idx >= len(lb.allFiles) {
-		return nil
-	}
-	fileRec := lb.allFiles[idx]
-
-	fileNumStr := fmt.Sprintf("%3d", idx+1)
-	name := fileRec.Filename
-	if r := []rune(name); len(r) > 12 {
-		name = string(r[:12])
-	}
-	fileNameStr := fmt.Sprintf("%-12s", name)
-	dateStr := fileRec.UploadedAt.Format("01/02/06")
-	sizeStr := fmt.Sprintf("%5s", compactFileSize(fileRec.Size))
-
-	markStr := " "
-	if lb.isFileTagged(fileRec.ID) {
-		markStr = "*"
-	}
-
-	line := lb.processedMidTemplate
-	line = strings.ReplaceAll(line, "^MARK", markStr)
-	line = strings.ReplaceAll(line, "^NUM", fileNumStr)
-	line = strings.ReplaceAll(line, "^NAME", fileNameStr)
-	line = strings.ReplaceAll(line, "^DATE", dateStr)
-	line = strings.ReplaceAll(line, "^SIZE", sizeStr)
-
-	// Build prefix for highlight rendering (ANSI-stripped).
-	prefixLine := strings.ReplaceAll(line, "^DESC", "")
-	processedPrefix := string(ansi.ReplacePipeCodes([]byte(prefixLine)))
-
-	dizLines := formatDIZLines(fileRec.Description, lb.descColWidth, dizMaxLines)
-
-	firstDesc := ""
-	if len(dizLines) > 0 {
-		firstDesc = dizLines[0]
-	}
-
-	var contLines []string
-	for i := 1; i < len(dizLines); i++ {
-		contLines = append(contLines, dizLines[i])
-	}
-
-	if len(contLines) > maxLines-1 {
-		contLines = contLines[:maxLines-1]
-	}
-
-	var result []string
-
-	if highlighted {
-		plainPrefix := lb.stripAnsi(processedPrefix)
-		plainDesc := lb.stripAnsi(firstDesc)
-		rowText := plainPrefix + plainDesc
-		visLen := ansi.VisibleLength(rowText)
-		if visLen < lb.termWidth {
-			rowText += strings.Repeat(" ", lb.termWidth-visLen)
-		}
-		result = append(result, lb.hiColorSeq+rowText+"\x1b[0m")
-	} else {
-		fullLine := strings.ReplaceAll(line, "^DESC", firstDesc)
-		processed := string(ansi.ReplacePipeCodes([]byte(fullLine)))
-		result = append(result, processed)
-	}
-
-	for _, cl := range contLines {
-		result = append(result, string(ansi.ReplacePipeCodes([]byte("|07"+lb.descIndentStr+cl))))
-	}
-
-	return result
-}
-
-func (lb *fileLightbar) screenRowForFile(fileIdx int) (startRow int, height int) {
-	if fileIdx < lb.topIndex {
-		return -1, 0
-	}
-	row := lb.fileAreaStartRow
-	for idx := lb.topIndex; idx < len(lb.allFiles) && (row-lb.fileAreaStartRow) < lb.visibleRows; idx++ {
-		h := lb.fileEntryHeight(idx)
-		remaining := lb.visibleRows - (row - lb.fileAreaStartRow)
-		if h > remaining {
-			h = remaining
-		}
-		if remaining < 1 {
-			break
-		}
-		if idx == fileIdx {
-			return row, h
-		}
-		row += h
-	}
-	return -1, 0
-}
-
-func (lb *fileLightbar) writeFileRow(screenRow int, fileIdx int, highlighted bool, height int) error {
-	lines := lb.buildFileEntry(fileIdx, highlighted, height)
-	for i, ln := range lines {
-		r := screenRow + i
-		if err := terminalio.WriteProcessedBytes(lb.terminal, []byte(ansi.MoveCursor(r, 1)+"\x1b[2K"), lb.outputMode); err != nil {
-			return err
-		}
-		if highlighted && i == 0 {
-			// First line already has raw ANSI from buildFileEntry; write directly.
-			if err := terminalio.WriteProcessedBytes(lb.terminal, []byte(ln), lb.outputMode); err != nil {
-				return err
-			}
-		} else if i == 0 {
-			if err := writeProcessedStringWithManualEncoding(lb.terminal, []byte(ln), lb.outputMode); err != nil {
-				return err
-			}
-		} else {
-			if err := terminalio.WriteProcessedBytes(lb.terminal, []byte(ln), lb.outputMode); err != nil {
-				return err
-			}
-		}
-	}
-	// Clear any remaining lines in this entry's allocated height.
-	for i := len(lines); i < height; i++ {
-		r := screenRow + i
-		if err := terminalio.WriteProcessedBytes(lb.terminal, []byte(ansi.MoveCursor(r, 1)+"\x1b[2K"), lb.outputMode); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (lb *fileLightbar) renderFileArea() error {
-	if err := terminalio.WriteProcessedBytes(lb.terminal, []byte(ansi.MoveCursor(lb.fileAreaStartRow, 1)), lb.outputMode); err != nil {
-		return err
-	}
-
-	linesUsed := 0
-	if len(lb.allFiles) == 0 {
-		msg := "|07   No files in this area."
-		if err := terminalio.WriteProcessedBytes(lb.terminal, []byte("\x1b[2K"), lb.outputMode); err != nil {
-			return err
-		}
-		if err := terminalio.WriteProcessedBytes(lb.terminal, ansi.ReplacePipeCodes([]byte(msg)), lb.outputMode); err != nil {
-			return err
-		}
-		linesUsed = 1
-	} else {
-		for idx := lb.topIndex; idx < len(lb.allFiles) && linesUsed < lb.visibleRows; idx++ {
-			h := lb.fileEntryHeight(idx)
-			remaining := lb.visibleRows - linesUsed
-			if h > remaining {
-				h = remaining
-			}
-			if remaining < 1 {
-				break
-			}
-			row := lb.fileAreaStartRow + linesUsed
-			if err := lb.writeFileRow(row, idx, idx == lb.selectedIndex, h); err != nil {
-				return err
-			}
-			linesUsed += h
-		}
-	}
-
-	// Clear unused rows.
-	for linesUsed < lb.visibleRows {
-		r := lb.fileAreaStartRow + linesUsed
-		if err := terminalio.WriteProcessedBytes(lb.terminal, []byte(ansi.MoveCursor(r, 1)+"\x1b[2K"), lb.outputMode); err != nil {
-			return err
-		}
-		linesUsed++
-	}
-	return nil
-}
-
-func (lb *fileLightbar) renderSeparator() error {
-	dashes := lb.termWidth - 2
-	if dashes < 0 {
-		dashes = 0
-	}
-	sep := "\xfa" + strings.Repeat("\xc4", dashes) + "\xfa"
-	sepLine := ansi.MoveCursor(lb.separatorRow, 1) + "\x1b[2K" + string(ansi.ReplacePipeCodes([]byte("|08"+sep+"|07")))
-	return terminalio.WriteProcessedBytes(lb.terminal, []byte(sepLine), lb.outputMode)
-}
-
-func (lb *fileLightbar) renderCmdBar() error {
-	if err := terminalio.WriteProcessedBytes(lb.terminal, []byte(ansi.MoveCursor(lb.cmdBarRow, 1)+"\x1b[2K"), lb.outputMode); err != nil {
-		return err
-	}
-	barWidth := 0
-	for ci, ent := range lb.cmdEntries {
-		barWidth += len(ent.label) + 2
-		if ci < len(lb.cmdEntries)-1 {
-			barWidth += 2
-		}
-	}
-	pad := (lb.termWidth - barWidth) / 2
-	if pad < 0 {
-		pad = 0
-	}
-	cmdBar := strings.Repeat(" ", pad)
-	for ci, ent := range lb.cmdEntries {
-		if ci == lb.cmdIndex {
-			cmdBar += ent.highlightColor + " " + ent.label + " " + "\x1b[0m"
-		} else {
-			cmdBar += ent.regularColor + " " + ent.label + " " + "\x1b[0m"
-		}
-		if ci < len(lb.cmdEntries)-1 {
-			cmdBar += "  "
-		}
-	}
-	return terminalio.WriteProcessedBytes(lb.terminal, []byte(cmdBar), lb.outputMode)
-}
-
-func (lb *fileLightbar) renderPageIndicator() error {
-	currentPage, calcTotalPages := lb.calculatePageInfo()
-
-	if len(lb.botContent) > 0 {
-		// Replace ^PAGE/^TOTALPAGES (legacy) and |FPAGE/|FTOTAL/@FPAGE@/@FTOTAL@ (new).
-		pageStr := string(processFileListPlaceholders([]byte(lb.botContent), currentPage, calcTotalPages, len(lb.allFiles), lb.fconfpath))
-		pageStr = strings.ReplaceAll(pageStr, "^PAGE", fmt.Sprintf("%d", currentPage))
-		pageStr = strings.ReplaceAll(pageStr, "^TOTALPAGES", fmt.Sprintf("%d", calcTotalPages))
-		processedPage := string(ansi.ReplacePipeCodes([]byte(pageStr)))
-		botLineSlice := strings.Split(processedPage, "\n")
-		for i, botLine := range botLineSlice {
-			botLine = strings.TrimRight(botLine, "\r")
-			plainLen := len(lb.stripAnsi(botLine))
-			linePad := (lb.termWidth - plainLen) / 2
-			if linePad < 0 {
-				linePad = 0
-			}
-			row := lb.termHeight - len(botLineSlice) + 1 + i
-			if row < 1 {
-				row = 1
-			}
-			centered := ansi.MoveCursor(row, 1) + "\x1b[2K" + strings.Repeat(" ", linePad) + botLine
-			if err := terminalio.WriteProcessedBytes(lb.terminal, []byte(centered), lb.outputMode); err != nil {
-				return err
-			}
-		}
-	}
-	return nil
-}
-
-func (lb *fileLightbar) renderTop() error {
-	if err := terminalio.WriteProcessedBytes(lb.terminal, []byte(ansi.MoveCursor(1, 1)), lb.outputMode); err != nil {
-		return err
-	}
-	curPage, calcTotalPages := lb.calculatePageInfo()
-	processed := ansi.ReplacePipeCodes(processFileListPlaceholders(lb.topTemplateBytes, curPage, calcTotalPages, len(lb.allFiles), lb.fconfpath))
-	return terminalio.WriteProcessedBytes(lb.terminal, processed, lb.outputMode)
-}
-
-func (lb *fileLightbar) renderFull() error {
-	if err := terminalio.WriteProcessedBytes(lb.terminal, []byte(ansi.ClearScreen()), lb.outputMode); err != nil {
-		return err
-	}
-	if err := lb.renderTop(); err != nil {
-		return err
-	}
-	if err := lb.renderFileArea(); err != nil {
-		return err
-	}
-	if err := lb.renderSeparator(); err != nil {
-		return err
-	}
-	if err := lb.renderCmdBar(); err != nil {
-		return err
-	}
-	return lb.renderPageIndicator()
 }
 
 func (lb *fileLightbar) run() (*user.User, string, error) {
@@ -1305,41 +827,4 @@ func (lb *fileLightbar) run() (*user.User, string, error) {
 			needFullRedraw = true
 		}
 	}
-}
-
-const (
-	dizMaxWidth = 45
-	dizMaxLines = 22
-)
-
-// formatDIZLines splits FILE_ID.DIZ content into display-ready lines.
-// Each line is truncated to maxWidth visible characters (ANSI-aware).
-// Returns at most maxLines lines, with trailing blank lines trimmed.
-func formatDIZLines(content string, maxWidth, maxLines int) []string {
-	if content == "" {
-		return nil
-	}
-
-	content = strings.ReplaceAll(content, "\r\n", "\n")
-	content = strings.ReplaceAll(content, "\r", "\n")
-
-	rawLines := strings.Split(content, "\n")
-
-	var lines []string
-	for _, line := range rawLines {
-		if len(lines) >= maxLines {
-			break
-		}
-		line = strings.TrimRight(line, " \t")
-		if ansi.VisibleLength(line) > maxWidth {
-			line = ansi.TruncateVisible(line, maxWidth)
-		}
-		lines = append(lines, line)
-	}
-
-	for len(lines) > 0 && strings.TrimSpace(lines[len(lines)-1]) == "" {
-		lines = lines[:len(lines)-1]
-	}
-
-	return lines
 }
